requirements.txt
djangorestframework 
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt